Move the Browser classes into the framework package.
This move eliminates the small browser package and results in a compact, complete framework package.
