+++ java-framework/src/main/java/examples/basic/TextFieldExample.java
@@ -1,12 +1,12 @@
 package examples.basic;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static webui.automation.browser.BrowserType.CHROME;
+import static webui.automation.framework.BrowserType.CHROME;
 
 import org.openqa.selenium.By;
 
-import webui.automation.browser.Browser;
 import webui.automation.elements.TextField;
+import webui.automation.framework.Browser;
 import webui.automation.framework.WebUI;
 
 /**
