--- java-framework/src/main/java/webui/automation/browser/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package defines the Browser Layer of the Web UI Automation Framework.
- */
-package webui.automation.browser;
