+++ java-framework/src/main/java/webui/automation/framework/BrowserType.java
@@ -1,4 +1,4 @@
-package webui.automation.browser;
+package webui.automation.framework;
 
 public enum BrowserType {
     CHROME, FIREFOX, IE
