+++ java-framework/src/main/java/webui/automation/framework/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package defines the core of the Web UI Automation Framework except the Browser layer.
+ * This package defines the core framework of the Web UI Automation Framework.
  */
 package webui.automation.framework;
